Add files via upload
This GitHub repository hosts the code and assets for a professionally designed and user-friendly website created for a fictional small business, which happens to be a coffee shop. The primary objective of this website is to establish a robust online presence for the coffee business, offering visitors a seamless experience while providing them with essential information and services.
The website comprises four key sections, each serving a distinct purpose. The homepage is designed to be visually appealing and informative, introducing visitors to the coffee business and showcasing its unique offerings. A dedicated Products Page provides comprehensive information about the business's products, including detailed descriptions, images, and pricing where applicable. The About Us Page tells the story of the coffee business, outlining its mission and values. Lastly, the Contact Page offers a platform for direct communication, featuring a user-friendly contact form for inquiries and providing the business's physical address and contact details.